requirement.txt
asgiref==3.2.5
blis==0.2.4
certifi==2019.11.28
chardet==3.0.4
ChatterBot==1.1.0
comtypes==1.1.7
cycler==0.10.0
cymem==2.0.3
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
en-core-web-sm==2.1.0
idna==2.9
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.2.1
mathparse==0.1.2
matplotlib==3.2.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.2
Pint==0.11
plac==0.9.6
preshed==2.0.1
PyAudio==0.2.11
pymongo==3.10.1
pyparsing==2.4.6
python-dateutil==2.7.5
pyttsx==1.1
pyttsx3==2.81
pytz==2019.3
pywin32==227
PyYAML==5.1.2
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
spacy==2.1.9
SpeechRecognition==3.8.1
SQLAlchemy==1.3.15
sqlparse==0.3.1
srsly==1.0.2
thinc==7.0.8
tqdm==4.43.0
urllib3==1.25.8
wasabi==0.6.0
